+++ src/codegen/java/edu/ie3/codegen/CopyBuilderGenerator.java
@@ -23,6 +23,7 @@ public final class CopyBuilderGenerator implements HelperMethods {
   private final ClassName builderClass;
   private final TypeVariableName builderTypeVariable = TypeVariableName.get("B");
   private final boolean hasParent;
+  private ModelDefinition parent;
   private ClassName parentBuilderClass;
 
   public CopyBuilderGenerator(
@@ -36,7 +37,7 @@ public final class CopyBuilderGenerator implements HelperMethods {
     this.hasParent = model.extendsName != null && !model.extendsName.isBlank();
 
     if (hasParent) {
-      ModelDefinition parent = getParent(model.extendsName, models);
+      this.parent = getParent(model.extendsName, models);
       this.parentBuilderClass = copyBuilderClassName(parent);
     }
   }
@@ -168,8 +169,13 @@ public final class CopyBuilderGenerator implements HelperMethods {
             .addModifiers(Modifier.PUBLIC, Modifier.STATIC);
 
     if (hasParent) {
-      TypeName parentBuilderType = ParameterizedTypeName.get(parentBuilderClass, builderClass);
-      builder.superclass(parentBuilderType);
+
+      if (parent.isAbstract) {
+        TypeName parentBuilderType = ParameterizedTypeName.get(parentBuilderClass, builderClass);
+        builder.superclass(parentBuilderType);
+      } else {
+        builder.superclass(parentBuilderClass);
+      }
     }
 
     // add the private fields to the copy builder