+++ src/main/java/edu/ie3/datamodel/models/input/connector/type/Transformer3WTypeInput.java
@@ -165,7 +165,7 @@ public class Transformer3WTypeInput extends AssetTypeInput {
    * @param tapNeutr Neutral tap position
    * @param tapMin Minimum available tap position
    * @param tapMax Maximum available tap position
-   * @param additionalInformation Of the input
+   * @param additionalInformation That were provided by the source
    */
   public Transformer3WTypeInput(
       UUID uuid,